python/requirements.txt
Flask==2.0.3
werkzeug==2.0.3 
qiskit==1.2.4 
qiskit-aer==0.15.1 
numpy==1.26.4
python-dotenv==0.19.2p